test(cdk/drag-drop): split standalone drag tests out into a separate file
Separates out the standalone dragging tests into a separate file to make them easier to navigate.
